Add support for customizing the return type of kd.call.
This approach is a bit ugly because it involves creating a value even though we're only going to use the type of that value, but it is nice in that it plays very nicely with the type inference model that we have.

I expect that most users will most likely only need to specify DataSlice and DataBag return types, hence the special support for those in py_boxing. Again, I was hesitating if adding a general boxing rule is too much for this, but it seems nice to avoid special cases for some operators/binding policies.

PiperOrigin-RevId: 680946567
Change-Id: Ib7fbe57c5895b629f8fe90eaadf3425a0e2ab064