Add the ability to specify return_type_as to apply_py, py_fn and bind.
I am not touching apply_py_on_cond and similar, since the filtering they do imply a DataSlice.

Working with functors that do not return DataSlice does get quite boilerplate-y (see test_bind_explicit_return_type for example). Still, I propose to check this in for now, and see how much it is used to decide if we need a more involved solution. For example, we could get rid of the need to specify this in py_fn and bind if we make functor evaluation pass the return type as a hidden input that the functor expression can use.

PiperOrigin-RevId: 681408854
Change-Id: I3d15e9534d74e985250054eae0bfaa35c4de8d3f