Add kde.core.uuids_with_allocation_size operator.
This operator creates a DataSlice whose items are Fingerprints of the seed and the desired size of the allocation.

This operator is useful for creating a DataSlice of distinct Uuids that should all share the same allocation id. Sharing the same allocation id can help to reduce the memory consumption when setting attributes on the objects.

PiperOrigin-RevId: 692902432
Change-Id: Ie3422977c25c220b7fb4779490e1233281203fd1