Introduce kd.named_schema.
For now the only benefits it provides versus kd.uu_schema is _not_ having an "attrs" argument, so there is no question whether the attributes affect the generated uuid. In the future, we are planning to also experiment with storing the name and including it in the repr. We are also planning to make schema='name' to be a shortcut for schema=kd.named_schema('name').

PiperOrigin-RevId: 694056743
Change-Id: I73a09920ee5c5f6af9069cc26dd5a12143c36061