Remove deprecation warning from as_itemid operator
These lambda operators are traced at import time rather than at runtime, meaning that the user will see the warning when importing Koda and not when evaluating the operator. 1) we shouldn't show a deprecation warning as the first thing the user sees when importing Koda and 2) it doesn't show when evaluating (or even binding) so it's rather pointless.

I've kept the deprecation warnings inside of the DataSliceView and DataSlice methods.

PiperOrigin-RevId: 696466665
Change-Id: I72211e0e6652fdfc080d60bcee9eaf041e9be59e