Implement eval_op for Koladata in C++
The goal is to provide an `eval_op` for Koladata with performance equivalent to `arolla.abc.aux_eval_op`, while supporting both deterministic and non-deterministic operators.

The main motivation for using C++ instead of Python is the need for access to the lower-level aux_bind_arguments, allowing us to avoid computing fingerprints for the input values.

Benchmarks:
```
_1 = kd.item(1)

%timeit arolla.abc.aux_eval_op('kde.add', _1, _1)   # 1.52us
%timeit expr_eval.eval_op('kde.add', _1, _1)        # 1.56us
%timeit expr_eval.eval_op('kde.list_shaped_as', _1) # 71.9us
```

PiperOrigin-RevId: 697625684
Change-Id: Ia4449e55248cc49a83342d4f8f89bb4d5f731bb8