Make error message optional in OperatorEvalError.
In this case we set the message from the status and don't add is a cause unless it has a payload. I am also removing "operator * failed during evaluation" part of the error messages. It takes space, but seems obvious. And if we decide to move operator names to the evaluation side, the new error will not look superficial.

PiperOrigin-RevId: 697942096
Change-Id: I9dd86060eaa093609d52c71f4edf16dae689715c