Add get_item as an overloadable operator
Subsequently reverses support for `db[ds]` directly inside of kd.get_item (introduced in cl/675495355) in favor of overloads instead. I believe this is a better way to do it, and it does not require us to bake unrelated functionality together into a single implementation. This is also better for extensions outside of core Koda that can overload existing ops.

In a follow-up change, I will use this to add `__getitem__` support for tuples.

PiperOrigin-RevId: 697979664
Change-Id: I399b8d18b7c85bf6e21d0a014626cbefdc01f9ec