Fix DataBag behavior inside `kde.from_py`:
- By default it doesn't create a databag;
- if multiple objects with different databags passed - it creates one for them.

See the new tests added for the examples where this change fixes the behavior.

Also move `from_py` implementation from a method of DataBag to the classmethod of DataSlice:

it does not necessarily return a DataBag any more (after this CL);
it is consistent with `DataSlice.from_vals` (and the behavior is similar in many cases).

PiperOrigin-RevId: 700340670
Change-Id: I203d947cc73c521034086fb64618757dc5351b7f