Unified binding policy
This patch introduces a new generic binding policy for Koladata.

The main motivation for creating a new policy, rather than modifying FullSignatureBindingPolicy is to avoid the need for migrating everything, while we still experiment with a different syntax for the operator declaration. Also, there is `_SelectBindingPolicy` subclass of `FullSignatureBindingPolicy` that needs special attention.

We modifying the signature coding, and embed the additional information about the parameters into the "aux_policy" string instead of default values. The string-based approach simplifies (and improves efficiency of) the processing, because we don't need to work with typed_values storing tuples.

It also leaves the default values in the signature to their original purpose to store the default values. E.g. we don't need to specify a default value if we call the operator using `bind_op`.

Preliminary benchmarks show the following performance improvements:
```
%timeit E.list_shaped_as(_1)    # 75.3us -> 8.22us
%timeit E.concat(_1, _1)        # 30.3us -> 8.57us

fn = kdf.fn(kde.add(I.x, I.y))
%timeit E.call(fn, x=_1, y=_1)  # 69.3us -> 25us
```
PiperOrigin-RevId: 701989863
Change-Id: I927c82e5c0767966d7fd6cdf54f786c0577d75ec