Add ability to pass argument schemas to kd.named_schema.
Initially we did not do it to avoid confusion with uu_schema which does take the argument schemas as input to the uuid generation, and named_schema that does not. But since they are named differently, I guess there is not much confusion after all, and this makes common usage of named_schema nicer.

PiperOrigin-RevId: 702308411
Change-Id: I8aeeacfd1eaf8b727046fbc88929a3a7ab2f58a9