Allow with_schema_from_obj on empty slices
Uses `agg_common_schema` which slightly changes the behavior for primitives, allowing values that are not strictly identical. Note that this is not required, and I can change this to a less aggressive approach (only allowing fully missing slices) instead if we prefer

Also shuffles around some operators to avoid dependency cycles and to allow infix operations (representing coalesce) to be used without worry.

PiperOrigin-RevId: 703513864
Change-Id: I655ca0603c372557e39a5876b2f670be19187788