Add a low level schema::CastDataTo
Why? In an upcoming change, I will remove the ability of CastToExplicit to cast ANY to an Entity schema. The proposed solution here is to not use ANY as a dummy schema, but to cast the data directly instead. This is safe since we produce the schema from the common schemas anyways, and we didn't do embedding before.

I've opted to not support embedding, as this muddies the purpose of the implementation (and the API becomes messy with 3 additional parameters that are largely unrelated). I suggest ToObject or the higher level casting utils are used instead. If we do need this in the future, we can revisit this decision.

PiperOrigin-RevId: 704634284
Change-Id: I428e2d31ed3c91bd373a0fb3bdf96efb60a36d9f