Disallow unsafe ToEntity casts
The current behavior is similar to a reinterpret cast, where no deep data validation is done. This can result in a bad state if the new schema doesn't match the underlying data. While explicit casts are allowed to be unsafe, they should either 1) succeed or 2) fail with a proper error message. We shouldn't allow bad state to form.

PiperOrigin-RevId: 704647859
Change-Id: I180ac0842034e04dbeabda1ad6c1e14d2cd63b84