Add the remaining Expectations to string operators.
I don't parse format string for the formatting operators, it seems to be too complex code for the improvement.

Performance wise we add a bit of `absl::StrCat`s to all the join / format operations (even if no error happens), but it seems to be significant only on small scalars (see benchmarks below), so I decided to go for it for the sake of code simplicity.

PiperOrigin-RevId: 711691212
Change-Id: Ifafee5bc3aa81238e7f8138337f947ecf00afba4