Allow slices as start/stop in kde.subslice.
This makes it an even more general building block for creating complex workflows. For example, a.expand_to(b, ndim=1) can now be implemented as a.S[kd.val_shaped_as(b, 0):].

The benchmark shows ~12% slowdown for the case where the start/stop are integers, which feels like not big enough to warrant a special fast path for that case.

PiperOrigin-RevId: 717850021
Change-Id: I33e35190af3c89e01256a4d58cd5f09705d44af1