Add the wrapped literal to the display name of kd.literal
If the debug repr is used, the literal currently shows an opaque name making introspection harder. This change aims to make it more clear which value is being shown in order to simplify debugging. Note that for e.g. `repr(kd.expr.literal(kd.slice(1)))`, the pretty print version is shown rather than the debug one, and is not affected by this change.

PiperOrigin-RevId: 718262183
Change-Id: I32d7e589c5839bc40db3ab96b8aa8a634fd1de4b