Change GetAttrOrMissing to use the schema to determine presence
Before this change (for example):
* `kd.new(x=1).with_schema(kd.schema.new_schema()).maybe('x')` -> `kd.slice(1)`
* `kd.new(x=kd.new(y=1)).with_schema(kd.schema.new_schema()).maybe('x')` -> ERROR
* `kd.new(x=kd.slice([1, 'a'])).with_schema(kd.schema.new_schema()).maybe('x')` -> `kd.slice([1, 'a']).as_any()`.

After this change:
* `kd.new(x=1).with_schema(kd.schema.new_schema()).maybe('x')` -> `kd.slice(None)`
* `kd.new(x=kd.new(y=1)).with_schema(kd.schema.new_schema()).maybe('x')` -> `kd.slice(None)`
* `kd.new(x=kd.slice([1, 'a'])).with_schema(kd.schema.new_schema()).maybe('x')` -> ``kd.slice([None, None])`.

Similarly, it's possible to get OBJECT lookups to fail if a `__schema__` is missing an attribute [1].

These are all examples of a case where the schema contains a subset of the data in the DataBag. There is no mismatch (unlike if the schema says INT32 and the data is FLOAT32), so we shouldn't fail or have inconsistent behavior for these cases. This CL attempts to remedy this by looking to the schema as the source of truth. If the schema indicates that there is no such attribute, we use None instead of attempting to look it up anyways. Crucially, we also do not fall back to ANY (which goes against the idea that ANY will only show if the user explicitly used it) so this CL is therefore also a step towards removing the ANY schema.

PiperOrigin-RevId: 723075887
Change-Id: I35a85f03cb0aef28d883118bac61986988341b08