Add support for customizing the tracing decorator behavior for "composite types".
Note that we do not add support for composite types for the top-level tracing (`kd.fn`), as functors can only return QValues now and kd.call can only accept QValues as arguments now. It seems that a trace_as_fn-decorated function is the right level of abstraction that allows us to insert the appropriate casts.

As a small bonus, now specifying "-> kd.types.DataBag" is a nicer alternative for "return_type_as=kd.types.DataBag".

PiperOrigin-RevId: 724355983
Change-Id: I318cab4c0c318a82811a57e711d254bc19018d6b