Allow NONE in DataSlice setters and return NONE in GetAttr for empty values
Examample: `kd.obj_like(kd.slice([])).not_really_missing_attr.get_schema()` was `OBJECT` and is now `NONE`.

I now allow NONE in GetDictItem, which fixes an issue where `kd.slice([]).with_bag(kd.bag())[0]` was previously not allowed, but now behaves like it had an OBJECT schema. As discussed in the chat, I also now (tangentially) allow setters to work on NONEs, since this is a no-op and NONEs can be implicitly casted to both Entities and OBJECTS.

Note that I'm hesitant to add NONE to e.g. IsListSchema (and same for Dicts etc), due to me interpreting them as "is exactly a list schema" rather than "can be interpreted as a list if casted".

PiperOrigin-RevId: 725999406
Change-Id: I7732f37fb5b5f2dc17cd248c9fc70d4e0ab75e46