Remove ANY (and friends) in python
I'm breaking up the removal in two stages (Python and C++) to make the changes more manageable to review. This CL removes access to ANY in Python, and since we have no ways to produce ANY unless the user explicitly requests it (at least once), it should not be possible to create unless you use custom C++ code. So while we do reduce the test coverage temporarily (until removing the C++ part as well), I believe this is fine.

PiperOrigin-RevId: 726433365
Change-Id: Icce2b1de10cf81ad184ebf39f8e4fdab3f08d7a6