Prepare DType for the removal of ANY
This CL changes how we compute the DTypeId from being generated from a list of supported types to being a fixed mapping. The reason is that we depend on the stability in several places: 1) the serialization of DType itself and 2) the stable fingerprint computed by kd.uuid. 1) can be addressed separately to avoid this dependency (e.g. using a proto enum), but 2 is more tricky. To allow ANY to be removed as soon as possible, I propose this solution - at least short term.

PiperOrigin-RevId: 728119087
Change-Id: Iaa72181c87627730d97d1ef65d287fed93756c81