Remove incorrect statement about schemas in kd.translate
The actual requirement is that `keys_to` must be narrow-castable to `keys_from`, which is non-trivial to explain and will likely lead to more confusion than it prevents. Since this is required essentially everywhere (e.g. assignment, binary operators, ...), and that this is already implied by the behavior of the operator, I think it's fine to leave unsaid.

PiperOrigin-RevId: 728708746
Change-Id: I01229e0e95bb46f1c74ac9e2a1678989ae96c326