Prevent `OperatorEvalError` from dropping non-Koda error payloads
Now we allow nesting non-Koda errors under Koda errors, but `OperatorEvalError` kept previous logic, which sometimes caused losing the payload. The side effect of the change is that some of the error `__cause__`s turn from `KodaError` to just `ValueError`. But it is no harm, because the main (raised) errors remain `KodaError`.

PiperOrigin-RevId: 730847716
Change-Id: I687be553e14f6a763dbc6661ab24cb02942a91f6