Add tests for kd.uuid stability
kd.uuid should be stable, but it's unclear as of now for how long we should guarantee stability (within the same process, between processes (short term), long term stability, ...). It's quite simple to accidentally break this stability (as I almost did when deprecating ANY - luckily we caught it before submitting). To prevent this, we test that some primitives, dtypes, ... remain stable as a sanity check.

PiperOrigin-RevId: 739973947
Change-Id: I9d87dfff8d8bff8a534bda6d12e5d481993c9a08