Remove unnecessary StatusOr from DataSlice::CreateFromScalar
CreateFromScalar will only succeed if `T` is supported, which is checked statically by `schema::GetDType<T>()`. The motivation here is that DataSlice::Create is surprisingly expensive so we should avoid it if possible. Also changes to take by value + std::move rather than const-ref since we will anyways create a copy inside of the DataItem constructor. This allows us to avoid copies in some cases which matters for e.g. Text.

PiperOrigin-RevId: 742268268
Change-Id: I891c17a83a161d4951f7493dc6e6d9b13afcdcd2