Reverse the dependency between functor and slices operators. As we add more complex functor operators such as loops, it is clear that functor should be at the top of the operator dependency tree.
This required making "internal_select" operator public, so adding tests for it and changing its API to take a Koda bool instead of Arolla bool.

This in turn led to some duplication of tests. I hope it is not a big deal. I have removed most tests for error messages, compared to select, since those make more sense on a user-facing operator.

PiperOrigin-RevId: 742300751
Change-Id: I69207800a10eafe501d0c6d17fe82bb52f4f5de4