A build script for PyPI packages for Arolla and Koladata
The packages are built together, so we have confidence in version compatibility.

We perform the build inside Docker using the ubuntu:22.04 image because it is relatively compatible with Google Colab.

The script is not intended as an ideal permanent solution, as minor changes are needed for virtually every release. For example, Google Colab recently migrated to Python 3.11 and upgraded to protobuf==5.29.

PiperOrigin-RevId: 747460624
Change-Id: I4eada39437c49b28babc42901e0721991097a92b