Truncate large strings to <=256 characters
This should make it more readable and will prevent crashes if the strings are super long. I've not added truncation to `str(ds)` to make it easier to escape the truncation (if this is wanted).

The change is made to the QValue only to avoid interfering with other usages of DataSliceRepr. LMK if you disagree with this decision.

PiperOrigin-RevId: 748577531
Change-Id: I7ca5bbfe88f641c6c43c90ca8ab38ae8aca35143