The simplest version of `kd.bind` operator implemented in C++.
I gave up on implementing that as a lambda operator. It seems that even if I would succeed --- the result will be too complex and extending will be quite difficult.

I didn't implement `_aux_fn_compute_variables` and co for now. I can do it as a follow up. I was not 100% sure that this use case is necessary for meta programming.

PiperOrigin-RevId: 748676753
Change-Id: Ife5ed17004b59f049eeda4f95c85737f2c2d0f10