Add the PersistedIncrementalDataBagManager abstraction.
It is often convenient to create a DataBag by incrementally adding smaller bags,
where each of the smaller bags is an update to the large DataBag.

This also provides the opportunity to persist the smaller bags separately.

Then at a later point, usually in a different process, one can reassemble the
large DataBag. But instead of loading the entire DataBag, one can load only the
smaller bags that are needed, thereby saving loading time and memory. In fact
the smaller bags can be loaded incrementally, so that decisions about which bags
to load can be made on the fly instead of up-front.

To streamline the consumption, you have to specify dependencies between the
smaller bags when they are added. It is trivial to specify a linear chain of
dependencies, but setting up a dependency DAG is easy and can significantly
improve the loading time and memory usage of data consumers.

The persistence of the bags happens by writing serialized bags and metadata to a filesystem directory. The directory is hermetic and can be copied/moved as desired.

PiperOrigin-RevId: 751406555
Change-Id: Iee12d893be1e61737af4d3752ef33191e8554721