Remove ownership from stream to stream writer
Motivation: To make the stream writer destructor automatically close the stream object, to prevent a situation where an owning callback waits on an orphan stream.

While we strongly advise that all streams be closed, it's better to be on the safe side. Particularly, if a computation was cancelled in the interactive mode, the code owning the writer may not receive control to properly close the stream.

PiperOrigin-RevId: 752663114
Change-Id: I9c72b845681afe397056433d692e375c78e4ff5d