Add support for empty indices in `__getitem__` with slices
This allows us to write generic code that always passes around DataSlices rather than having to deal specifically with `None`.

The CPython code has been adapted to handle this as well. The subscript method could likely be removed in favor of the operator, but since the assignment subscript method doesn't have a corresponding operator (since it mutates the data), we need the logic anyways. Hence, I've kept the subscript method as well.

PiperOrigin-RevId: 755306428
Change-Id: I1ce7c3554b7c70a0fffd545069989225eb0738c3