[RFC] Truncate the middle of DataSlice strings rather than the end
For truncation length 6:
* Before: `kd.slice(['abcdefghi')` -> `DataSlice(['abcdef...'])`
* After: `kd.slice(['abcdefghi')` -> `DataSlice(['abc'...'ghi'])`

Please let me know if you have other suggestions, such as `'abc' + ... + 'ijk'`, `'abcdef'...`, `'abcdef'<...>` or `'abcdef' (truncated)`. Similarly, for bytestrings, I've only kept the `b` prefix on the first part. If we do e.g. `b'abc' + ... + b'ijk'`, it may be better to include it for both parts.

The benefits of this approach is that 1) it's more obvious that truncation has happened (which is currently a user concern), as the ellipsis is not part of the string, and 2) one sees both the start and end of the string which likely displays are richer picture of the string itself.

PiperOrigin-RevId: 756213521
Change-Id: I2fb74f56bb988656c2d7fc40f3a7fad37bbe5dd3