Change executor.schedule and stream.subscribe_once to forward the cancellation context, but not automatically react to it
This makes the behaviour closer to the C++.

Motivation: While having a built-in support for the cancellation within the scheduling mechanism seems like a good idea to enforce consistency, its behaviour looks surprising.

  event = threading.Event()
  reader.subscribe_once(executor, event.set)
  event.wait()

If the current context gets cancelled, the event may never be set and the thread will wait forever. And it might look unexpected that closing of the stream corresponding to the reader will not help, because although it would trigger scheduling of the callback on the executor, the executor would silently drop the task because the current context is cancelled.

---

Possible alternatives / follow-ups:
 * generate a warning when the executor drops a task (cons: this may cause noise in the interactive mode)
 * provide compatible synchronisation primitives (cons: we currently don't have a need for them, so preferably we should use the standard primitives provided by Python)
 * provide a "read_all" / "wait" method for the stream reader, so there is no need for the "advanced" API (possible follow-up).

PiperOrigin-RevId: 757208969
Change-Id: I2980ed355e52c18daae3de9d6168df38aade9ac7