Utilities to propagate Python stack trace frame through C++.
On Python side we are using a `DataSlice` with three fields, on C++ side we use a simple struct and add it as an error payload. Alternatively we could have put this `DataSlice` straight into the struct. But we still need the code to read its fields somewhere & there is a chance that we will want to move `StackTraceFrame` to Arolla in future.

PiperOrigin-RevId: 758776502
Change-Id: I24f91d67a23649ad7fa4c1bdd283ec66bc1edbc3