Add ExpectDType
In the end, I decided to keep the existing ExpectString etc for easy use in C++. For ExpectNumeric and ExpectInteger, I've kept the specific implementation since it indeed reads weirdly to see "expected values of float64" when you actually just require them to be numerics (like you mentioned offline).

I've also opted to keep the error message the same as before and do not include "castable to" or similar. This makes this change minimal and I slightly prefer it over the more verbose version. LMK if you disagree

PiperOrigin-RevId: 761390480
Change-Id: I164484d395f27fec10131189d449a4334417b7c8