[RFC] Change kd.assertion.with_assertion to accept a callback functor
This allows us to implement expensive error message construction without paying for it upfront.

I want to keep the implementation in `assertion.py` to be able to use it from other operators (such as `kd.disjoint_coalesce`) even though we support functors. Due to this, the QExpr implementation is done in the functor directory. It's a bit of a messy compromise. LMK what you think of this.

Another option is to keep kd.assertion.with_assertion as-is and add another operator specifically for the callback. However, I think don't really see the point and it may harm discovery.

PiperOrigin-RevId: 765080322
Change-Id: I6b21f11d96a3834c800f97cc7eba1c3f1677d28c