Closing a stream_writer makes Orphaned() return true
This change generalizes the behaviour of the `Orphaned()` method. Previously, `Orphaned()` only returned true when no potential readers were left. Now, it means the writer can no longer make visible effects on any reader.

Motivation: This simplifies the implementation of various streaming operations by allowing the closing of the resulting stream to serve as a flag to signal that the process should stop (either because of completion or interruption).
PiperOrigin-RevId: 765787393
Change-Id: Id1d5caf977ee7080d714bbd3739d65a51bd8dd4d