Address a leak detector error for mock objects in basic_routine_test
What happened? In the previous implementation, the gMock framework sometimes detected that the mock object was not deleted after the tests had finished (even though the test itself contained a check that the mock's destructor had been called). Our assumption is that this was happening because of some form of a race condition.

The fix has been tested using `bazel test --runs_per_test_detects_flakes --runs_per_test=1000` on the `CancellationContextPropagation` test.

PiperOrigin-RevId: 766596759
Change-Id: I68346e1d6d72a3c3fe4da11ef0ccf66f3b09892f