[RFC] Rename kd.tuple to kd.tuples and drop "make_" prefix from tuple ops
Initially, I wanted to just create aliases in order to avoid a breaking change. However, if we want to have `kd.tuple` be the same as `kd.make_tuple` we need to change the `kd.tuple` namespace to `kd.tuples` thus introducing a breaking change. IMO, we can then just as well break things a bit more to have a nice state.

PiperOrigin-RevId: 770712202
Change-Id: I36124a8bcf8c4df296aff7858d768048d38d483f