Add a DataSlicePath abstraction to help with handling persisted incremental data.
A data slice path is a sequence of actions to perform on a DataSlice to obtain
a subslice. The set of allowed actions is small and simple:
* Getting the keys of a dict.
* Getting the values of a dict.
* Exploding a list.
* Getting an attribute of an entity.

The utility of a data slice path is that it specifies how to obtain a subslice
without doing it immediately. Thus, we have the possibility to gather more
information about the context of the subslice, for example, the schemas of its
ancestor data slices.

PiperOrigin-RevId: 775213448
Change-Id: Ic37fcc14d4c392010b4b2eee640dee1a9ae619df