Ban the use of kd.OBJECT schema in the schema_helper for persisted incremental data.
Much of the benefit of indexing incremental updates with nodes in the schema graph would be lost if we allowed kd.OBJECT. Banning it seems reasonable for ensuring good performance.

It does affect the ability to accommodate Koda objects such as functors, but supporting them inside persisted incremental dataslices is not a high priority.

PiperOrigin-RevId: 777533276
Change-Id: I83de62151d4a03d6ca0049f90aaf32042b95bbe3