Support operators returning (but not consuming!) an iterable in the default transformations, by converting iterable to an everything-available-at-once stream.
This has a danger that somebody implementing a new operation will implement only the version returning an iterable, and forget that they also need the stream version and the conversion rule, but hopefully this risk is low since it should be clear that if you only create an iterable and not a stream in a custom C++ op, there's no way for the computation to be actually streaming.

The automatic conversion in the other direction, from a stream to a future[iterable] is much more dangerous, as it can lead to "collapsing" of streams that are in fact already meaningfully streaming, so I think we should not do it.

PiperOrigin-RevId: 778891389
Change-Id: I2adcf9e5ada761415619d11fcdaee9b4124e3543