ProtoDescriptorFromSchema: convert DICT schemas correctly into protobuf map fields.
Protobuf map fields have special requirements because of legacy reasons (compatibility with proto2). As a result, a map field "my_map" must be associated with an entry message whose name is "MyMapEntry" (fixed in this change), with "key" and "value" fields, and if the "value" is a message, then it must exist on the same or higher level as "MyMapEntry" (fixed in this change).

PiperOrigin-RevId: 783259655
Change-Id: I1f1e433fb3222c08e4e8e8945a38fee67484a8bf