Minor improvement on `value_counts`, by using `collapse` instead of `unique`.
While the implementation using `unique` is correct, it depends on the consistent handling of unique and agg_count on sort=False. By using the grouped in both dict key and value, we are confident that the implementation is correct, without depending on the handling details of `unique` and `group_by`.

PiperOrigin-RevId: 785326985
Change-Id: I5ba9e1c5ba53d1dd0e117cbe32c3bc526e9dca7d