Fix bug in DataSliceImpl serialization.
The bug was happening if we serialize DataSlice with a huge string followed by empty strings till the end of the slice. If the last extra proto contains only empty strings, it has estimated_size = 0 and wasn't stored to the output. To fix it we make estimated sizes of an empty string to be one byte.

PiperOrigin-RevId: 785437523
Change-Id: If7d0eaf66601fe44c3c7fa7e931108a8b8d795e4