Make sure that all nodes specified as extra_nodes_to_extract are guaranteed to become variables.
Previously this was failing for literal DataSlices, which was breaking assumptions in the parallel evaluation code.

This does make the generated functors a bit more ugly for the "named 1+-dim DataSlice literal" case because of an extra aux variable for the result of the explode, but I think it is a reasonable price to pay to avoid handling named nodes and extra_nodes_to_extract differently.

PiperOrigin-RevId: 786218762
Change-Id: If4d5f86ddccdfc604453e051d813b151b2bfbb57