Add TraceMe to backend operators
Changes ReturnsOperatorEvalError into a more generic KodaOperatorWrapper that does the operator handling plus profiling. This allows us to inject this functionality without adding a new wrapper (similar to ReturnsOperatorEvalError).

Eventually, we may wish to move this logic into the Arolla backend - but this is more complicated so I hesitate to do so right now.

PiperOrigin-RevId: 788556754
Change-Id: I72d791be16502786482f27cfae8650b1764d05a0