Add `DataSliceAction.get_subschema_bag()`.
Associating each action with the minimal schema bag that facilitates the action is an important step to represent the schema incrementally in `schema_helper.py`. The goal is to have a method `SchemaHelper.get_schema_bag(schema_node_names)` that assembles a minimal schema for the given `schema_node_names`. That method will be called by `PersistedIncrementalDataSliceManager`, which will attach the minimal schema to the minimal dataslice that was requested by the user.

PiperOrigin-RevId: 791586450
Change-Id: I63bf72365bf9f54270ebcfb7a806369a3628e460