Add a SimpleInMemoryDataSliceManager, which is used to test the consistency of managers with vanilla Koda.
Created DataSliceManagerInterface, which is implemented by SimpleInMemoryDataSliceManager and PersistedIncrementalDataSliceManager.

Refactored a test to make the difference between vanilla Koda and PersistedIncrementalDataSliceManager clear when an itemid is associated with more than one schema. For that reason, it states in the docstring of update() that the behavior is undefined if an itemid is associated with two or more structured schemas.

PiperOrigin-RevId: 792066591
Change-Id: If1e8ce8b465101b22eda9832f002952a534fcd73