Simplify auto_variables logic by always extracting nodes passed as extra_nodes_to_extract, even if they already exist as a variable.
Previously, we were relying on "is_shared_node" to save the day when a certain expression that we need to extract is present both as its own variable, and as part of another variable, as written in the comment I'm deleting. However, is_shared_node logic is disabled for "simple" nodes such as I.x, so we had a bug in that case.

Rather than trying to fix that bug, I propose to get rid of this complication, as it only affects the readability of functors created with non-empty extra_nodes_to_extract which is not as important.

PiperOrigin-RevId: 797712135
Change-Id: I3e2f3fd2ecba7afd6cf55331aca80560a569399c