Initialize schema node to data bag mapping with empty lists.
In the past we initialized it with singleton lists that contained the initial bag name ''. But that is not needed anymore - the new implementation of update() can already associate some schema node names with the empty list. The dependency on the initial empty bag is already implicit, and _persist_schema_node_name_to_data_bag_names() already handles the case when the list is empty.

PiperOrigin-RevId: 799983542
Change-Id: Ie5635d80358fa225a1e6f80a1556c0a605cd0be8