Add `with_attrs` as a method to extension types
I'm a bit on the fence about adding it as a method. On the one hand, it makes it more straightforward to use - especially for consumers of specific extension types that may not necessarily know it's an extension type. On the other hand, it prevents users from defining their own `with_attrs` method. But given that we also inject `eval` and similar, maybe this is fine.

PiperOrigin-RevId: 801837477
Change-Id: I9072bb981e113bdc8e64698156b58aade075fbf4