Refactor `extension_type_registry.dynamic_cast` to be eager-only
This clarifies the responsibility of each part: the one defined in kde is a lazy op and the one defined in extension_type_registry is an eager op. I plan to follow the same pattern for future operators.

Also stops relying on self annotations for virtual methods. Ideally, Functor annotations should only cause upcasting so it feels bad to rely on it to also support downcasting. Inside of the virtual methods, we actually do want to downcast so here it's more clear to do it explicitly.

PiperOrigin-RevId: 802102504
Change-Id: Ia1ab244db1a3eadf05b01a48416b4452513d1111