Add an argument `action_history_index` to `PersistedIncrementalDataSliceManager.branch()`.
This new parameter allows branching a `PersistedIncrementalDataSliceManager` from a specific point in its action history. The branched manager will start with the state of the original manager after the action at the given index was performed.

This provides a simple rollback mechanism: you create a branch that starts from a rolled-back state. You can add updates to that branch, and of course updates that happened after `action_history_index` are still visible in the main manager's current/latest state. So the rollback doesn't affect the main manager - it is only visible in the branch.

PiperOrigin-RevId: 802487080
Change-Id: If4b3fb4a967d61de4c0cf39e4cb42966581a8ffa