Implement transactional updates in `PersistedIncrementalDataBagManager`.
This change introduces a `metadata_update_number` to the metadata proto and modifies how metadata is persisted. Each metadata update is written to a versioned file (e.g., `metadata-000000000001.pb`) using a file rename operation that can be atomic (if the implementation supports it). This ensures that only one writer can successfully commit a metadata update at a time, preventing data corruption from concurrent writes.

This change also ensures that the cached metadata is consistent with the metadata on disk, and that interrupts (keyboard interrupts by users, out of disk space errors, etc.) cannot lead to managers with inconsistent states.

PiperOrigin-RevId: 804872381
Change-Id: I60866945edaac36820fa1468e6522b091b4aaf99