Make updates of `PersistedIncrementalDataSliceManager` transactional (atomic).
This change introduces a `metadata_update_number` to the metadata proto and uses versioned filenames for metadata files (e.g., `metadata-000000000001.pb`). Metadata updates are now performed by writing to a temporary file and then atomically renaming it to the versioned filename with `overwrite=False`. This allows for detecting and preventing concurrent writes, ensuring data consistency, when file-to-file renaming is atomic.

This change also ensures that the cached metadata is consistent with the metadata on disk, and that interrupts (keyboard interrupts by users, out of disk space errors, etc.) cannot lead to managers with inconsistent states (e.g. manager._schema_helper is out of sync with the schema from the metadata).

The SliceManager now always passes an explicit set of bag names to the BagManager instances. I.e. there is no more code that relies on whatever happens to be loaded in a BagManager. That way, one or more BagManagers can commit updates and the SliceManager won't rely on them until it committed its transaction.

PiperOrigin-RevId: 805326481
Change-Id: Id24acac7bfb5b22523c22cf0b26bf9654c415ede