Add support for null extension types
This is added both in the form of operators and as an optional mixin class that adds methods. Using NullableMixin has zero cost unless you are actually making it null. I chose the mixin API instead of 1) always supporting this or 2) adding e.g. `@kd.extension_type(nullable=True)` because:

1. It forces you to think about nullability if you want it. This makes it more likely that you remember to deal with it, and also makes the API clearer for the user.
2. The mixin makes the inheritance more clear, while kd.extension_type(nullable=True) would require it to be specified for all types in the hierarchy and would also require injections into the original class (not just the QValue and ExprViews).

Let me know if you disagree with this.

PiperOrigin-RevId: 805368572
Change-Id: I3de6552032d717d532f5b7bad97dcf342c222d11