PersistedIncrementalDataSliceManager: use uuids for DataBag names.
Before this change, it used timestamps with microsecond granularity. The microsecond granularity, and the reliance on the system time, which can be changed/reset and updated via NTP, means that collisions were possible. Conflicting bag names would be rejected by a PersistedIncrementalBagManager, so there was never any possibility for data corruption.

However, UUID is much closer semantically to what we want here.

Since PersistedIncrementalDataSliceManager now keeps a timestamped history, no information is lost by this change.

PiperOrigin-RevId: 805793416
Change-Id: Id4a79db929709c9e19c72e51f8b6fb0b1e2a1821