Don't assume calls to `_get_fresh_bag_name()` will return lexicographically increasing names.
It happens to be the case that `uuid.uuid1()` returns increasing UUIDs, and before that the code used timestamps for bag names, which also increased.

But the code should sort explicitly when it needs a list of sorted names.

Also updated the test to remove an assertion that would fail if `_get_fresh_bag_name()` did not return lex increasing names. The canonical topological sorting agrees with the order in which the bags were added to the bag manager. If the names of the bags are not increasing, then the canonical topological sorting won't be lex increasing either.

PiperOrigin-RevId: 806262030
Change-Id: I623183819b83fa19c7f60789eb42f66e69c22380